Resize charts and move resources_rc.py
