+++ Right Posture/bin/Data/save_setting.json
@@ -1,1 +1,1 @@
-{"Light": 0, "DND": 0, "Discord": 0, "PreCam1": 1, "PreDetail": 1, "PreLog": 1}
+{"Light": 0, "DND": 0, "Discord": 0, "PreCam1": 0, "PreDetail": 1, "PreLog": 1}
