+++ Right Posture/main.ui
@@ -2133,7 +2133,7 @@ p, li { white-space: pre-wrap; }
                            <number>10</number>
                           </property>
                           <property name="bottomMargin">
-                           <number>10</number>
+                           <number>12</number>
                           </property>
                           <item>
                            <widget class="QLabel" name="labelBoxBlenderInstalation_8">
@@ -2239,7 +2239,7 @@ p, li { white-space: pre-wrap; }
                            <number>6</number>
                           </property>
                           <property name="bottomMargin">
-                           <number>10</number>
+                           <number>0</number>
                           </property>
                           <item>
                            <widget class="QFrame" name="frame_donut">
