+++ Right Posture/modules/ui_login.py
@@ -19,7 +19,7 @@ from PySide6.QtWidgets import (QApplication, QFrame, QGridLayout, QHBoxLayout,
     QLabel, QLineEdit, QMainWindow, QPushButton,
     QSizePolicy, QSpacerItem, QStackedWidget, QVBoxLayout,
     QWidget)
-from .resources_rc import *
+import resources_rc
 
 class Ui_Login(object):
     def setupUi(self, Login):
